Chapter 07 Programming Projects: Modify sum.c to sum double values

+++ chapter_06/sum.c
@@ -1,19 +1,20 @@
 // Write a program that takes a series of numbers and sums them together.
+// Chapter 07, modify the program so that it sums two double values.
 
 #include <stdio.h>
 
 int main(void)
 {
 
-    long num, sum = 0;
+    double num, sum = 0;
 
     while (num != 0)
     {
         printf("Please enter a number: ");
-        scanf("%ld", &num);
+        scanf("%le", &num);
 
         sum += num;
-        printf("Your current sum: %ld\n", sum);
+        printf("Your current sum: %f\n", sum);
     }
 
     return 0;
